fix(loop): support 32-bit monotonic ticks and animations beyond MAX_BLOCK_SYMBOLS

+++ include/scyan/types.h
@@ -78,7 +78,7 @@ struct custom_status_state {
     bool is_idle;
     uint8_t bongo_state; // 0: idle/neutral, 1: tap left, 2: tap right
     uint8_t wpm_tick;    // Increments on each time interval tick for scrolling WPM chart
-    uint8_t loop_tick;   // Increments on animation frame interval for Loop widgets
+    uint32_t loop_tick;  // Increments on animation frame interval for Loop widgets
 };
 
 #ifndef WIDGET_TYPE_BONGO
