+++ src/engine.h
@@ -18,5 +18,5 @@ void engine_set_idle(bool idle);
 void engine_bongo_tap(bool is_left);
 uint8_t engine_get_bongo_state(void);
 uint8_t engine_get_wpm_tick(void);
-uint8_t engine_get_loop_tick(void);
+uint32_t engine_get_loop_tick(void);
